requirements.txt
attrs==21.2.0
Brotli==1.1.0
ConfigParser==6.0.0
cryptography==3.4.8
Cython==3.0.8
dagshub==0.3.9
Django==4.2.7
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
importlib_metadata==6.8.0
importlib_metadata==4.6.4
ipython==8.19.0
ipywidgets==7.8.1
Jinja2==3.1.2
jnius==1.1.0
keyring==23.5.0
lockfile==0.12.2
lxml==4.9.4
nltk==3.6.7
numpy==1.26.3
ordereddict==1.1
pandas==2.2.0
Pillow==10.1.0
Pillow==9.0.1
Pillow==10.2.0
protobuf==3.12.4
psycopg2_binary==2.9.9
pycocotools==2.0.7
pyOpenSSL==24.0.0
python-dotenv==1.0.1
railroad==0.5.0
Sphinx==7.2.6
thread==0.1.3
toml==0.10.2
torch==2.1.0+cpu
torchvision==0.16.0+cpu
tornado==6.4
tqdm==4.66.1
xmlrpclib==1.0.1
zenml==0.55.1
zipp==1.0.0
